Small changes to the UI.

+++ src/Camera.cpp
@@ -14,7 +14,7 @@ Camera::Camera(const unsigned int _width, const unsigned int _height,
         fov(_fov), near(_near), far(_far),
         acceleration(_acceleration)
 {
-    aspect = width / height;
+    aspect = (float)width / height;
     pitch = yaw = speed = 0.f;
 }
